Set the far plane in the Camera's GetProjectionMatrix function to 1000f which fixes the issue of the surrounding cube not fully drawing.
Created a debug camera, which is basically just a first person camera that can move around freely. Also experimented with zooming. The debug camera can zoom in and out with the scroll wheel, and can be reset by pressing the middle mouse button. Movement is regular WASD controls, with spacebar and left shift being used to go up and down the y-axis.

To switch to the debug camera, press the tilde key (~)